set isRoot as false for mac package build
