+++ internal/gui/gui_linux.go
@@ -5,6 +5,7 @@ package gui
 
 import (
 	"os"
+	"os/user"
 )
 
 // TODO (darkrengarius): change path
@@ -16,4 +17,14 @@ const (
 func checkIsPackage() bool {
 	_, err := os.Stat(deinstallerPath)
 	return err == nil
+}
+
+func isRoot() bool {
+	userLvl, err := user.Current()
+	if err == nil {
+		if userLvl.Username == "root" {
+			return true
+		}
+	}
+	return false
 }
